ecommerce-tests/e2e/requirements.txt
pytest==7.3.1
requests==2.29.0
pytest-html==3.2.0
pytest-vcr==1.0.2
pytest-asyncio==0.21.0
logging-formatter-anticrlf==1.2.0
python-dotenv==1.0.0
